requirements.txt
charset-normalizer==2.1.0
coreapi==2.3.3
Django==5.0
django-extensions==3.1.5
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
djangorestframework==3.15.1
django-filter
drf-yasg==1.21.14
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary>=2.9.10
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
setuptools>=70.0.0
django-redis==5.4.0
celery==5.3.6
numpy==2.4.2
djangorestframework-simplejwt==5.5.1